+++ src/main/java/org/silsagusi/joonggaemoa/domain/article/controller/ComplexController.java
@@ -1,0 +1,26 @@
+package org.silsagusi.joonggaemoa.domain.article.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.silsagusi.joonggaemoa.domain.article.service.ArticleResponse;
+import org.silsagusi.joonggaemoa.domain.article.service.ComplexResponse;
+import org.silsagusi.joonggaemoa.domain.article.service.ComplexService;
+import org.silsagusi.joonggaemoa.global.api.ApiResponse;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+@RestController
+@RequiredArgsConstructor
+public class ComplexController {
+
+    private final ComplexService complexService;
+
+    @GetMapping("/api/naverbudongsan/complexes")
+    public ResponseEntity<ApiResponse<List<ComplexResponse>>> getComplexes() {
+        List<ComplexResponse> complexResponseList = complexService.getComplex().stream()
+                .map(ComplexResponse::of).toList();
+        return ResponseEntity.ok(ApiResponse.ok(complexResponseList));
+    }
+}
