--- src/main/java/org/silsagusi/joonggaemoa/domain/crawler/naverbudongsan/repository/ArticleRepository.java
@@ -1,7 +1,0 @@
-package org.silsagusi.joonggaemoa.domain.crawler.naverbudongsan.repository;
-
-import org.silsagusi.joonggaemoa.domain.crawler.naverbudongsan.entity.Article;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ArticleRepository extends JpaRepository<Article, Long> {
-}
