--- src/main/java/org/silsagusi/joonggaemoa/domain/crawler/naverbudongsan/repository/ComplexRepository.java
@@ -1,7 +1,0 @@
-package org.silsagusi.joonggaemoa.domain.crawler.naverbudongsan.repository;
-
-import org.silsagusi.joonggaemoa.domain.crawler.naverbudongsan.entity.Complex;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ComplexRepository extends JpaRepository<Complex, Long> {
-}
