+++ src/main/java/org/silsagusi/joonggaemoa/request/naverland/client/dto/ClientRegionRequestHeader.java
@@ -1,0 +1,19 @@
+package org.silsagusi.joonggaemoa.request.naverland.client.dto;
+
+public class ClientRegionRequestHeader {
+    public String rletTpCd = "APT:OPST:VL:ABYG:OBYG:JGC:JWJT:DDDGG:SGJT:HOJT:JGB:OR:GSW:SG:SMS:GJCG:GM:TJ:APTHGJ";
+
+    public String tradTpCd = "A1:B1:B2:B3";
+
+    public String latitude;
+
+    public String longitude;
+
+    public final String[] cortar = {
+            "1168010100",
+            "1168010300",
+            "1168010400",
+            "1168010500",
+            "1168000600"
+    };
+}
