+++ src/main/java/org/silsagusi/joonggaemoa/request/naverland/client/dto/ClientRegionResponse.java
@@ -1,0 +1,26 @@
+package org.silsagusi.joonggaemoa.request.naverland.client.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class ClientRegionResponse {
+
+    // 법정동 코드
+    private String cortarNo;
+
+    // 화면 정가운데 위도
+    private Double centerLat;
+
+    // 화면 정가운데 경도
+    private Double centerLon;
+
+    // 법정동명
+    private String cortarName;
+
+    // 법정동 타입 (city 시/도, dvsn (division) 시/군/구, sec (section) 읍/면/동)
+    private String cortarType;
+}
